refactor(chat): reorganize chat code for better clarity and consistency
- Moved code from 'chatService.ts' and some code from 'ChatDataProvider.ts' into 'chatRenderer.ts', 'chatMessageHandler.ts', and 'utils.ts'.
- Removed 'chatService.ts' after migrating its functionality to new files in src/views/chatView.
- Moved chat command registration from 'extension.ts' to 'registry.ts' and 'cmds.ts' in src/commands/chat.
- Added a usage warning message to the open chat button.

Improves overall code clarity and structure.
