+++ src/commands/chat/registry.ts
@@ -1,0 +1,34 @@
+// Copyright(c) ZenML GmbH 2024. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0(the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied.See the License for the specific language governing
+// permissions and limitations under the License.
+import { chatCommands } from './cmds';
+import { registerCommand } from '../../common/vscodeapi';
+import { ZenExtension } from '../../services/ZenExtension';
+import { ExtensionContext, commands } from 'vscode';
+
+export const registerChatCommands = (context: ExtensionContext) => {
+  try {
+    const registeredCommands = [
+      registerCommand('zenml.openChat', () => chatCommands.openChat(context)),
+    ];
+
+    registeredCommands.forEach(cmd => {
+      context.subscriptions.push(cmd);
+      ZenExtension.commandDisposables.push(cmd);
+    });
+
+    commands.executeCommand('setContext', 'chatCommandsRegistered', true);
+  } catch (error) {
+    console.error('Error registering chat commands:', error);
+    commands.executeCommand('setContext', 'chatCommandsRegistered', false);
+  }
+};
