+++ src/services/ZenExtension.ts
@@ -44,7 +44,7 @@ import { toggleCommands } from '../utils/global';
 import { PanelDataProvider } from '../views/panel/panelView/PanelDataProvider';
 import { ComponentDataProvider } from '../views/activityBar/componentView/ComponentDataProvider';
 import { registerComponentCommands } from '../commands/components/registry';
-
+import { registerChatCommands } from '../commands/chat/registry';
 export interface IServerInfo {
   name: string;
   module: string;
@@ -75,6 +75,7 @@ export class ZenExtension {
     registerComponentCommands,
     registerPipelineCommands,
     registerSecretsCommands,
+    registerChatCommands,
   ];
 
   /**
