+++ src/views/chatView/chatMessageHandler.ts
@@ -1,0 +1,23 @@
+// Copyright(c) ZenML GmbH 2024. All Rights Reserved.
+// Licensed under the Apache License, Version 2.0(the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at:
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied.See the License for the specific language governing
+// permissions and limitations under the License.
+import { ChatDataProvider } from './ChatDataProvider';
+
+export async function handleWebviewMessage(message: any, chatDataProvider: ChatDataProvider) {
+  if (message.command === 'sendMessage' && message.text) {
+    await chatDataProvider.addMessage(message.text, message.context);
+  }
+
+  if (message.command === 'clearChat') {
+    await chatDataProvider.clearChatLog();
+  }
+}
